Add Content Collection to project
Content Collection is an Astro feature that allows stores collections of content within a unique  directory within an Astro project. This allows easy management and thus where all dummy content is placed for this project. This also includes a config file which allows customization of content as well as utilize typescript features for the content